move old data file
