+++ lat/generate_steering_vectors.py
@@ -4,7 +4,7 @@ import os
 from tqdm import tqdm
 from dotenv import load_dotenv
 from lat.modeling import Llama7BChatHelper
-from lat.data import ComparisonDataset
+from lat.data.data import ComparisonDataset
 from lat.utils import system_prompt, data_path
